refactor: Split crawler.py into separate CLI and class modules
